src/requirements.txt
django
psycopg[binary]
djangorestframework
djangorestframework-simplejwt
django-otp
pyotp
drf-yasg